Move from nonbonded to utils
